Introduce base adapter class
